Added 2 new apis for homepage and counts

+++ src/main/java/com/tool/ProjectTool/model/response/NoteList.java
@@ -1,18 +1,20 @@
 package com.tool.ProjectTool.model.response;
 
+import java.math.BigInteger;
+
 public class NoteList {
 	
-	private Integer id;
+	private BigInteger id;
 	
 	private String title;
 	
 	private String description;
 
-	public Integer getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
