+++ src/main/java/com/tool/ProjectTool/service/NotesService.java
@@ -1,5 +1,6 @@
 package com.tool.ProjectTool.service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Iterator;
@@ -50,7 +51,7 @@ public class NotesService {
 			Object[] row = (Object[]) itr.next();
 			NoteList note = new NoteList();
 
-			note.setId((Integer) row[0]);
+			note.setId((BigInteger) row[0]);
 			note.setDescription(String.valueOf(row[1]));
 			note.setTitle(String.valueOf(row[2]));
